+++ src/Core/Managed/Shared/DataContracts/DependencyTelemetry.cs
@@ -131,32 +131,12 @@ namespace Microsoft.ApplicationInsights.DataContracts
         }
 
         /// <summary>
-        /// Gets or sets request count. Obsolete, use SamplingPercentage instead.
-        /// </summary>
-        [Obsolete]
-        [EditorBrowsable(EditorBrowsableState.Never)]
-        public int? Count
-        {
-            get { return 1; }
-            set { }
-        }
-
-        /// <summary>
         /// Gets or sets a value indicating whether the dependency call was successful or not.
         /// </summary>
         public override bool? Success
         {
             get { return this.Data.success; }
             set { this.Data.success = value; }
-        }
-
-        /// <summary>
-        /// Gets or sets a value indicating whether the dependency call was made asynchronously or not.
-        /// </summary>
-        public bool? Async
-        {
-            get { return this.Data.async; }
-            set { this.Data.async = value; }
         }
 
         /// <summary>
@@ -180,22 +160,8 @@ namespace Microsoft.ApplicationInsights.DataContracts
 
             set
             {
-                if (string.IsNullOrEmpty(value))
-                {
-                    this.Data.dependencyKind = BondDependencyKind.Other;
-                }
-                else
-                {
-                    // There is no TryParse in .Net 3.5
-                    if (Enum.GetNames(typeof(BondDependencyKind)).Contains(value))
-                    {
-                        this.Data.dependencyKind = (BondDependencyKind)Enum.Parse(typeof(BondDependencyKind), value);
-                    }
-                    else
-                    {
-                        this.Data.dependencyKind = BondDependencyKind.Other;
-                    }
-                }
+                BondDependencyKind result;
+                this.Data.dependencyKind = Enum.TryParse(value, true, out result) ? result : BondDependencyKind.Other;
             }
         }