+++ src/Core/Managed/Shared/Extensibility/Implementation/TransmissionProcessor.cs
@@ -5,6 +5,7 @@
     using System.Text;
     using Microsoft.ApplicationInsights.Channel;
     using Microsoft.ApplicationInsights.Extensibility.Implementation.Tracing;
+    using Platform;
 
     /// <summary>
     /// An <see cref="ITelemetryProcessor"/> that act as a proxy to the Transmission of telemetry"/>.
@@ -13,7 +14,8 @@
     /// </summary>
     internal class TransmissionProcessor : ITelemetryProcessor
     {        
-        private readonly TelemetryConfiguration configuration;        
+        private readonly TelemetryConfiguration configuration;
+        private readonly IDebugOutput debugOutput;     
 
         /// <summary>
         /// Initializes a new instance of the <see cref="TransmissionProcessor"/> class.
@@ -27,6 +29,7 @@
             }
 
             this.configuration = configuration;
+            this.debugOutput = PlatformSingleton.Current.GetDebugOutput();
         }
 
         /// <summary>
@@ -41,6 +44,8 @@
             }
 
             this.configuration.TelemetryChannel.Send(item);
+
+            TelemetryDebugWriter.WriteTelemetry(item);
         }
     }
 }